added titledQueryEditorInput
