+++ src/sql/workbench/test/browser/parts/editor/editorStatusModeSelect.test.ts
@@ -24,7 +24,7 @@ import { UntitledTextEditorInput } from 'vs/workbench/services/untitled/common/u
 import { IUntitledTextEditorService } from 'vs/workbench/services/untitled/common/untitledTextEditorService';
 import { FileEditorInput } from 'vs/workbench/contrib/files/common/editors/fileEditorInput';
 import { URI } from 'vs/base/common/uri';
-import { FileQueryEditorInput } from 'sql/workbench/services/query/common/fileQueryEditorInput';
+import { FileQueryEditorInput } from 'sql/workbench/contrib/query/common/fileQueryEditorInput';
 import { QueryResultsInput } from 'sql/workbench/common/editor/query/queryResultsInput';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
 import { EditorType } from 'vs/editor/common/editorCommon';
